hackathon/utils/computer_vision.py
import cv2
import numpy as np
from PIL import Image
import io
import random
import os

# In a production environment, we would load the actual models
# For this demo, we'll simulate computer vision functionality
# Note: TensorFlow import is commented out to avoid compatibility issues
# import tensorflow as tf

def load_object_detection_model():
    """
    Load a pre-trained object detection model.
    In a real implementation, this would load an actual TensorFlow or PyTorch model.
    """
    # This is a placeholder for model loading
    # In production, we would have code like:
    # model = tf.saved_model.load('models/object_detection_model')
    
    print("Object detection model loaded")
    return "object_detection_model"

def load_spoilage_detection_model():
    """
    Load a pre-trained food spoilage detection model.
    In a real implementation, this would load an actual model.
    """
    # This is a placeholder for model loading
    # In production, we would have code like:
    # model = tf.keras.models.load_model('models/spoilage_detection_model')
    
    print("Spoilage detection model loaded")
    return "spoilage_detection_model"

def preprocess_image(image_bytes):
    """
    Preprocess the image for model input.
    
    Args:
        image_bytes: Raw image bytes from uploaded file
        
    Returns:
        Image processed for model input
    """
    # Convert bytes to numpy array
    nparr = np.frombuffer(image_bytes, np.uint8)
    image = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
    
    # Process image (resize, normalize, etc.)
    image = cv2.resize(image, (640, 480))
    image = image / 255.0  # Normalize
    
    return image

def detect_objects_in_image(image_bytes):
    """
    Detect objects in an uploaded image.
    
    Args:
        image_bytes: Raw image bytes from uploaded file
        
    Returns:
        List of detected items with confidence scores
    """
    # In a real implementation, we would:
    # 1. Preprocess the image
    # 2. Run it through a pre-trained model
    # 3. Process the results
    
    # For this demo, we'll return simulated results
    common_kitchen_items = [
        {"item": "Tomatoes", "quantity": 5, "confidence": 0.98, "expiry_days": 4},
        {"item": "Lettuce", "quantity": 2, "confidence": 0.95, "expiry_days": 3},
        {"item": "Chicken", "quantity": 3, "confidence": 0.92, "expiry_days": 2},
        {"item": "Onions", "quantity": 8, "confidence": 0.97, "expiry_days": 7},
        {"item": "Bell Peppers", "quantity": 4, "confidence": 0.93, "expiry_days": 5},
        {"item": "Carrots", "quantity": 6, "confidence": 0.96, "expiry_days": 10},
        {"item": "Potatoes", "quantity": 10, "confidence": 0.98, "expiry_days": 14},
        {"item": "Herbs (Basil)", "quantity": 1, "confidence": 0.87, "expiry_days": 2},
        {"item": "Lemons", "quantity": 3, "confidence": 0.91, "expiry_days": 8},
        {"item": "Milk", "quantity": 1, "confidence": 0.94, "expiry_days": 5}
    ]
    
    # Simulate detection by randomly selecting a subset of items
    num_items = random.randint(4, 8)
    detected_items = random.sample(common_kitchen_items, num_items)
    
    # Add slight randomness to confidence scores
    for item in detected_items:
        item['confidence'] = min(1.0, item['confidence'] + random.uniform(-0.05, 0.05))
    
    return detected_items

def detect_food_spoilage(image_bytes):
    """
    Detect potential food spoilage in an uploaded image.
    
    Args:
        image_bytes: Raw image bytes from uploaded file
        
    Returns:
        List of potentially spoiled items with confidence scores
    """
    # In a real implementation, we would:
    # 1. Preprocess the image
    # 2. Run it through a specialized food spoilage model
    # 3. Process the results
    
    # For this demo, we'll return simulated results
    spoilage_candidates = [
        {"item": "Lettuce", "confidence": 0.78, "issue": "Wilting signs"},
        {"item": "Tomatoes", "confidence": 0.65, "issue": "Early mold signs"},
        {"item": "Herbs (Basil)", "confidence": 0.88, "issue": "Browning leaves"},
        {"item": "Chicken", "confidence": 0.72, "issue": "Color change"}
    ]
    
    # Simulate spoilage detection
    num_spoiled = random.randint(0, 2)  # Most times, we'll have 0-2 spoiled items
    if num_spoiled == 0:
        return []
    
    spoiled_items = random.sample(spoilage_candidates, num_spoiled)
    
    # Add slight randomness to confidence scores
    for item in spoiled_items:
        item['confidence'] = min(1.0, item['confidence'] + random.uniform(-0.05, 0.05))
    
    return spoiled_items
