hackathon/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
    "statsmodels>=0.14.4",
    "streamlit>=1.44.0",
    "tensorflow>=2.14.0",
]
